Gamepad teleop only allow forward motion once docked
When using gamepad teleop to dock the robot on a
docking station, the left joystick could be used to
push the robot back against the docking station or
turn in-place while docked. These motions from the
robot can disconnect or damage the station. In this
commit, once the robot detects charge, the in-place
rotation and backwards motions are disabled.

Additionally, this commit tries to counteract the
noise made by the wheel steppers when holding the
robot against the docking station by putting the
wheels into freewheel mode.