Remove docked motion limiter
This logic intended to prevent rotational/reverse motion in
gamepad teleop when the robot is docked on the docking station.
While it succeeds, it also prevents these motions while the
robot is plugged into a charger (not a station). Since the
robot has no way to differentiate the two at a low level, I'm
removing this logic for now. Some design thinking will be
needed before this feature can be reintroduced.