Pinocchio requires lower numpy version
Using the latest 1.x version of numpy (1.26.4) is not possible
with Pinoochio, a dependency of Web Teleop, because it depends
on cmeel-boost, which depends on numpy 1.24.x. In a previous
PR, we increased the numpy version from 1.23.x to make trimesh
happy. To keep both happy, this commit pins on >=1.24